% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCIT_package.R
\docType{package}
\name{CCIT-package}
\alias{CCIT-package}
\alias{CCIT}
\title{The 'CCIT' package.}
\description{
TBD: Add description
}
\references{
TBD: Add reference.
}
\author{
Naeem KHoshnevis

Nina Katz-Kristy

Falco Joannes Bargagli Stoffi
}
