% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create.sequence.R
\name{create.sequence}
\alias{create.sequence}
\title{TBD}
\usage{
create.sequence(exploration.dat, ulist.used, parms.used)
}
\arguments{
\item{exploration.dat:}{exploration dataframe with outcome Y and exposure
variable treat}

\item{ulist.used:}{user-specified functions to supply to rpart
(return value of define_u.list)}

\item{parms.used:}{additional parameters to supply to rpart}
}
\value{
list of nested decision trees and corresponding goodness metrics
}
\description{
Calculates the sequence of candidate trees.
}
