% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpart_funcs.R
\name{define_u.list}
\alias{define_u.list}
\title{TBD}
\usage{
define_u.list(stopping.rule = TRUE)
}
\arguments{
\item{stopping.rule:}{A boolean to indicate whether the tree-splitting
algorithm should stop when the estimated interaction effect is <1/10 of
the overall effect.}
}
\description{
Defines functions for custom rpart function

required libraries: rpart, rpart.plot
}
