% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data_covs.R
\name{generate_syn_data_covs}
\alias{generate_syn_data_covs}
\title{Generate Synthetic Data for CausalGPS Package - Modified by Nina}
\usage{
generate_syn_data_covs(sample_size = 10000, gps_spec = 1)
}
\description{
Generates synthetic data set based on different GPS models and covariates.

Required Libraries: truncnorm, dplyr
}
