% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data_het.R
\name{generate_syn_data_het}
\alias{generate_syn_data_het}
\title{Generate Synthetic Data for CausalGPS Package - Modified by Nina}
\usage{
generate_syn_data_het(
  sample_size = 10000,
  outcome_type = "continuous",
  outcome_sd = 1,
  gps_spec = 1,
  cova_spec = 1,
  em_spec = 1,
  heterogenous_intercept = FALSE,
  em_as_confounder = FALSE,
  beta = 1
)
}
\arguments{
\item{sample_size}{Number of data samples.}

\item{outcome_type}{continuous or binary outcome}

\item{outcome_sd}{Standard deviation used to generate the outcome in the synthetic data set}

\item{gps_spec}{A numerical value (1-7) that indicates the GPS model
used to generate synthetic data. See the code for more details.}

\item{cova_spec}{A numerical value (1-2) to modify the covariates. If
cova_spec = 1, covariates are mutually independent. See the
code for more details.}

\item{em_spec}{A numerical value (1-2) that indicates the effect modification model.
em_spec = 1 -> separate main effects of effect modifiers on treatment effect
(i.e. no interaction between effect modifiers)
em_spec = 2 -> interaction between effect modifiers}

\item{heterogenous_intercept}{Allow intercepts to differ by effect modifier level}

\item{em_as_confounder}{Effect modifier 1 has linear association with treatment
(i.e. is a confounder)}
}
\value{
\code{synthetic_data}: The function returns a data.frame saved the
constructed synthetic data.
}
\description{
Generates synthetic data set based on different GPS models and covariates.
}
\examples{

set.seed(298)
s_data <- generate_syn_data(sample_size=100,
                                  outcome_sd = 10, gps_spec = 1,
                                  cova_spec = 1)

Required Libraries: truncnorm, dplyr
}
