% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data_outcome.R
\name{generate_syn_data_outcome}
\alias{generate_syn_data_outcome}
\title{Generate Synthetic Data for CausalGPS Package - Modified by Nina}
\usage{
generate_syn_data_outcome(
  cf1,
  cf2,
  cf3,
  cf4,
  cf5,
  cf6,
  em1,
  em2,
  em3,
  em4,
  treat,
  outcome_sd = 1,
  em_spec = 1,
  heterogenous_intercept = FALSE,
  beta = 1
)
}
\description{
Generates synthetic data set based on different GPS models and covariates.

Required Libraries: truncnorm, dplyr
}
