% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dataset.R
\name{split_dataset}
\alias{split_dataset}
\title{Split data set}
\usage{
split_dataset(data, num_exposure_cats)
}
\arguments{
\item{data:}{A data.frame with variable "treat" corresponding to
exposure level}

\item{num_exposure_cats:}{the number of categories to bin the exposure
level into for stratification}
}
\description{
Splits data set into training, validation, and inference, stratifying
by exposure level.

Required Libraries: caret, dplyr
}
