% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stratified_GPS_matching.R
\name{stratified_GPS_matching}
\alias{stratified_GPS_matching}
\title{Stratify GPS matching}
\usage{
stratified_GPS_matching(
  data,
  delta_n,
  exposure_name,
  confounders_names,
  names_of_strata_vars,
  outcome_name
)
}
\arguments{
\item{data:}{A data frame with data to be matched.}

\item{delta_n:}{A bin width parameter to pass to the CausalGPS package.}

\item{exposure_name:}{The name of the exposure variable.}

\item{confounders_names:}{A vector of strings representing names of the
confounding variables.}

\item{names_of_strata_vars:}{A vector of strings representing names of
variables to stratify by.}

\item{outcome_name:}{The name of the outcome variable.}
}
\description{
TBD
}
