% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_synthetic_data_outcome.R
\name{generate_syn_data_outcome}
\alias{generate_syn_data_outcome}
\title{Generate outcome for synthetic data}
\usage{
generate_syn_data_outcome(
  cf,
  em,
  treat,
  outcome_sd = 1,
  em_spec = 1,
  heterogenous_intercept = FALSE,
  beta = 1
)
}
\arguments{
\item{cf}{A data.frame of confounder variables}

\item{em}{A data.frame of effect modifier variables}

\item{treat:}{A vector of treatment variable}

\item{outcome_sd:}{standard deviation of outcome variable}

\item{em_spec:}{specification of effect modifier function
(possible values: 0, 1, 2)
\itemize{
\item em_spec = 0: no effect modification
\item em_spec = 1:independent effect modification of em1 and em2 (4 causal effect groups)
\item em_spec = 2 -> interactive effect modification of em1 and em2 (3 causal effect groups)
}}

\item{heterogenous_intercept:}{whether the intercept of the CRF should vary by effect modifier levels}

\item{beta:}{strength of effect modification (only effective if em_spec is 1 or 2)}
}
\description{
Generates synthetic outcome variable given covariate, effect modifier, and
treatment variables.
}
